+++ src/base.rs
@@ -1,8 +1,10 @@
+use failure::ResultExt;
 use j4rs::Instance;
 use j4rs::InvocationArg;
 use j4rs::Jvm;
 use serde::de::DeserializeOwned;
 
+use super::ErrorKind;
 use super::MBeanInfo;
 use super::Result;
 
@@ -35,11 +37,12 @@ impl MBeanAddress {
                     "service:jmx:rmi://{}/jndi/rmi://{}/jmxrmi",
                     address, address
                 );
-                jvm.create_instance(JMX_SERVICE_URL, &vec![InvocationArg::from(url)])?
+                jvm.create_instance(JMX_SERVICE_URL, &vec![InvocationArg::from(url)])
+                    .with_context(|_| ErrorKind::JavaCreateInstance(JMX_SERVICE_URL))?
             },
             MBeanAddress::ServiceUrl(url) => jvm.create_instance(
                 JMX_SERVICE_URL, &vec![InvocationArg::from(url)]
-            )?,
+            ).with_context(|_| ErrorKind::JavaCreateInstance(JMX_SERVICE_URL))?,
         };
         Ok(instance)
     }