+++ src/lib.rs
@@ -1,7 +1,5 @@
-#[macro_use]
-extern crate error_chain;
+extern crate failure;
 extern crate j4rs;
-
 extern crate serde;
 #[macro_use]
 extern crate serde_derive;
@@ -9,16 +7,15 @@ extern crate serde_derive;
 
 mod base;
 mod constants;
-mod errors;
+mod error;
 mod mbean_client;
 mod mbean_info;
 mod util;
 
 
-pub use self::errors::Error;
-pub use self::errors::ErrorKind;
-pub use self::errors::Result;
-pub use self::errors::ResultExt;
+pub use self::error::Error;
+pub use self::error::ErrorKind;
+pub use self::error::Result;
 
 pub use self::base::MBeanAddress;
 pub use self::base::MBeanClientTrait;