+++ src/mbean_client.rs
@@ -1,9 +1,11 @@
+use failure::ResultExt;
 use j4rs::Instance;
 use j4rs::InvocationArg;
 use j4rs::Jvm;
-use j4rs::new_jvm;
+use j4rs::JvmBuilder;
 use serde::de::DeserializeOwned;
 
+use super::ErrorKind;
 use super::MBeanAddress;
 use super::MBeanClientTrait;
 use super::MBeanInfo;
@@ -40,7 +42,7 @@ impl MBeanClient {
     pub fn connect_with_options(
         address: MBeanAddress, options: MBeanClientOptions
     ) -> Result<MBeanClient> {
-        let jvm = MBeanClient::into_jvm(options.jvm)?;
+        let jvm = options.jvm.build().with_context(|_| ErrorKind::JvmInit)?;
         let service_url = address.for_java(&jvm)?;
         MBeanClient::connect_service_url(jvm, service_url)
     }
@@ -60,17 +62,10 @@ impl MBeanClient {
 
     /// Helper to obtain a `javax.management.MBeanServerConnection` instance.
     fn get_connection(jvm: &Jvm, server: &Instance) -> Result<Instance> {
-        let connection = jvm.invoke(server, "getMBeanServerConnection", &vec![])?;
+        let connection = jvm.invoke(server, "getMBeanServerConnection", &vec![]).with_context(
+            |_| ErrorKind::JavaInvoke(server.class_name().to_string(), "getMBeanServerConnection")
+        )?;
         Ok(connection)
-    }
-
-    /// Create a default JVM if none was provided.
-    fn into_jvm(jvm: Option<Jvm>) -> Result<Jvm> {
-        if let Some(jvm) = jvm {
-            Ok(jvm)
-        } else {
-            Ok(new_jvm(vec![], vec![])?)
-        }
     }
 
     /// Helper to obtain a `javax.management.remote.JMXConnectorFactory` instance.
@@ -78,7 +73,7 @@ impl MBeanClient {
         let server = jvm.invoke_static(
             JMX_CONNECTOR_FACTORY, "connect",
             &vec![InvocationArg::from(service_url)]
-        )?;
+        ).with_context(|_| ErrorKind::JavaInvokeStatic(JMX_CONNECTOR_FACTORY, "connect"))?;
         Ok(server)
     }
 }
@@ -91,12 +86,15 @@ impl MBeanClientTrait for MBeanClient {
     {
         let object_name = self.jvm.create_instance(
             JMX_OBJECT_NAME, &vec![InvocationArg::from(mbean.into())]
-        )?;
+        ).with_context(|_| ErrorKind::JavaCreateInstance(JMX_OBJECT_NAME))?;
         let value = self.jvm.invoke(
             &self.connection, "getAttribute",
             &vec![InvocationArg::from(object_name), InvocationArg::from(attribute.into())]
+        ).with_context(
+            |_| ErrorKind::JavaInvoke(self.connection.class_name().to_string(), "getAttribute")
         )?;
-        Ok(self.jvm.to_rust(value)?)
+        let value: T = self.jvm.to_rust(value).with_context(|_| ErrorKind::RustCast("<dynamic>"))?;
+        Ok(value)
     }
 
     fn get_mbean_info<S>(&self, mbean: S) -> Result<MBeanInfo>
@@ -104,10 +102,12 @@ impl MBeanClientTrait for MBeanClient {
     {
         let object_name = self.jvm.create_instance(
             JMX_OBJECT_NAME, &vec![InvocationArg::from(mbean.into())]
-        )?;
+        ).with_context(|_| ErrorKind::JavaCreateInstance(JMX_OBJECT_NAME))?;
         let info = self.jvm.invoke(
             &self.connection, "getMBeanInfo",
             &vec![InvocationArg::from(object_name)]
+        ).with_context(
+            |_| ErrorKind::JavaInvoke(self.connection.class_name().to_string(), "getMBeanInfo")
         )?;
         MBeanInfo::from_instance(&self.jvm, info)
     }
@@ -118,21 +118,29 @@ impl MBeanClientTrait for MBeanClient {
     {
         let name = self.jvm.create_instance(
             JMX_OBJECT_NAME, &vec![InvocationArg::from(name.into())]
-        )?;
+        ).with_context(|_| ErrorKind::JavaCreateInstance(JMX_OBJECT_NAME))?;
         let query = self.jvm.create_instance(
             JMX_OBJECT_NAME, &vec![InvocationArg::from(query.into())]
-        )?;
-        let query = self.jvm.cast(&query, JMX_QUERY_EXP)?;
+        ).with_context(|_| ErrorKind::JavaCreateInstance(JMX_OBJECT_NAME))?;
+        let query = self.jvm.cast(&query, JMX_QUERY_EXP)
+            .with_context(|_| ErrorKind::JavaCast(JMX_QUERY_EXP.to_string()))?;
         let names = self.jvm.invoke(
             &self.connection, "queryNames",
             &vec![InvocationArg::from(name), InvocationArg::from(query)]
+        ).with_context(
+            |_| ErrorKind::JavaInvoke(self.connection.class_name().to_string(), "queryNames")
         )?;
-        let names = self.jvm.invoke(&names, "toArray", &vec![])?;
+        let names = self.jvm.invoke(&names, "toArray", &vec![]).with_context(
+            |_| ErrorKind::JavaInvoke(names.class_name().to_string(), "toArray")
+        )?;
         let names = to_vec(&self.jvm, names, JMX_OBJECT_NAME)?;
         let mut result = Vec::new();
         for instance in names {
-            let instance = self.jvm.invoke(&instance, "toString", &vec![])?;
-            let name: String = self.jvm.to_rust(instance)?;
+            let instance = self.jvm.invoke(&instance, "toString", &vec![]).with_context(
+                |_| ErrorKind::JavaInvoke(instance.class_name().to_string(), "toString")
+            )?;
+            let name: String = self.jvm.to_rust(instance)
+                .with_context(|_| ErrorKind::RustCast("String"))?;
             result.push(name);
         }
         Ok(result)
@@ -141,23 +149,22 @@ impl MBeanClientTrait for MBeanClient {
 
 
 /// Additional `MBeanClient` connection options.
-#[derive(Clone)]
-pub struct MBeanClientOptions {
-    jvm: Option<Jvm>,
+pub struct MBeanClientOptions<'a> {
+    jvm: JvmBuilder<'a>,
 }
 
-impl MBeanClientOptions {
-    /// Use the given JVM instance instead of creating one.
-    pub fn jvm(mut self, jvm: Jvm) -> Self {
-        self.jvm = Some(jvm);
+impl<'a> MBeanClientOptions<'a> {
+    /// Use the given JvmBuilder instance instead of the default one.
+    pub fn builder(mut self, builder: JvmBuilder<'a>) -> Self {
+        self.jvm = builder;
         self
     }
 }
 
-impl Default for MBeanClientOptions {
+impl<'a> Default for MBeanClientOptions<'a> {
     fn default() -> Self {
         MBeanClientOptions {
-            jvm: None,
+            jvm: JvmBuilder::new(),
         }
     }
 }