+++ src/mbean_info.rs
@@ -1,6 +1,8 @@
+use failure::ResultExt;
 use j4rs::Instance;
 use j4rs::Jvm;
 
+use super::ErrorKind;
 use super::Result;
 
 use super::constants::JMX_MBEAN_ATTRIBUTE_INFO;
@@ -27,21 +29,39 @@ impl MBeanAttribute {
     /// Create an `MBeanAttribute` instance from a `javax.management.MBeanAttributeInfo`
     /// java instance.
     pub fn from_instance(jvm: &Jvm, instance: Instance) -> Result<MBeanAttribute> {
-        let is_is = jvm.invoke(&instance, "isIs", &vec![])?;
-        let is_is: bool = jvm.to_rust(is_is)?;
-        let is_readable = jvm.invoke(&instance, "isReadable", &vec![])?;
-        let is_readable: bool = jvm.to_rust(is_readable)?;
-        let is_writable = jvm.invoke(&instance, "isWritable", &vec![])?;
-        let is_writable: bool = jvm.to_rust(is_writable)?;
-        let type_name = jvm.invoke(&instance, "getType", &vec![])?;
-        let type_name: String = jvm.to_rust(type_name)?;
+        let is_is = jvm.invoke(&instance, "isIs", &vec![]).with_context(
+            |_| ErrorKind::JavaInvoke(instance.class_name().to_string(), "isIs")
+        )?;
+        let is_is: bool = jvm.to_rust(is_is)
+            .with_context(|_| ErrorKind::RustCast("bool"))?;
+        let is_readable = jvm.invoke(&instance, "isReadable", &vec![]).with_context(
+            |_| ErrorKind::JavaInvoke(instance.class_name().to_string(), "isReadable")
+        )?;
+        let is_readable: bool = jvm.to_rust(is_readable)
+            .with_context(|_| ErrorKind::RustCast("bool"))?;
+        let is_writable = jvm.invoke(&instance, "isWritable", &vec![]).with_context(
+            |_| ErrorKind::JavaInvoke(instance.class_name().to_string(), "isWritable")
+        )?;
+        let is_writable: bool = jvm.to_rust(is_writable)
+            .with_context(|_| ErrorKind::RustCast("bool"))?;
+        let type_name = jvm.invoke(&instance, "getType", &vec![]).with_context(
+            |_| ErrorKind::JavaInvoke(instance.class_name().to_string(), "getType")
+        )?;
+        let type_name: String = jvm.to_rust(type_name)
+            .with_context(|_| ErrorKind::RustCast("String"))?;
 
         // Cast to javax.management.MBeanFeatureInfo for the other attributes.
-        let instance = jvm.cast(&instance, JMX_MBEAN_FEATURE_INFO)?;
-        let description = jvm.invoke(&instance, "getDescription", &vec![])?;
-        let description: String = jvm.to_rust(description)?;
-        let name = jvm.invoke(&instance, "getName", &vec![])?;
-        let name: String = jvm.to_rust(name)?;
+        let instance = jvm.cast(&instance, JMX_MBEAN_FEATURE_INFO)
+            .with_context(|_| ErrorKind::JavaCast(JMX_MBEAN_FEATURE_INFO.to_string()))?;
+        let description = jvm.invoke(&instance, "getDescription", &vec![]).with_context(
+            |_| ErrorKind::JavaInvoke(instance.class_name().to_string(), "getDescription")
+        )?;
+        let description: String = jvm.to_rust(description)
+            .with_context(|_| ErrorKind::RustCast("String"))?;
+        let name = jvm.invoke(&instance, "getName", &vec![]).with_context(
+            |_| ErrorKind::JavaInvoke(instance.class_name().to_string(), "getName")
+        )?;
+        let name: String = jvm.to_rust(name).with_context(|_| ErrorKind::RustCast("String"))?;
         Ok(MBeanAttribute {
             description,
             is_is,
@@ -72,10 +92,16 @@ impl MBeanInfo {
     /// Create an `MBeanInfo` instance from a `javax.management.MBeanInfo` java instance.
     pub fn from_instance(jvm: &Jvm, instance: Instance) -> Result<MBeanInfo> {
         let attributes = MBeanInfo::attributes_from_instance(jvm, &instance)?;
-        let class_name = jvm.invoke(&instance, "getClassName", &vec![])?;
-        let class_name: String = jvm.to_rust(class_name)?;
-        let description = jvm.invoke(&instance, "getDescription", &vec![])?;
-        let description: String = jvm.to_rust(description)?;
+        let class_name = jvm.invoke(&instance, "getClassName", &vec![]).with_context(
+            |_| ErrorKind::JavaInvoke(instance.class_name().to_string(), "getClassName")
+        )?;
+        let class_name: String = jvm.to_rust(class_name)
+            .with_context(|_| ErrorKind::RustCast("String"))?;
+        let description = jvm.invoke(&instance, "getDescription", &vec![]).with_context(
+            |_| ErrorKind::JavaInvoke(instance.class_name().to_string(), "getDescription")
+        )?;
+        let description: String = jvm.to_rust(description)
+            .with_context(|_| ErrorKind::RustCast("String"))?;
         Ok(MBeanInfo {
             attributes,
             class_name,
@@ -86,7 +112,9 @@ impl MBeanInfo {
 
 impl MBeanInfo {
     fn attributes_from_instance(jvm: &Jvm, instance: &Instance) -> Result<Vec<MBeanAttribute>> {
-        let raw = jvm.invoke(instance, "getAttributes", &vec![])?;
+        let raw = jvm.invoke(instance, "getAttributes", &vec![]).with_context(
+            |_| ErrorKind::JavaInvoke(instance.class_name().to_string(), "getAttributes")
+        )?;
         let raw = to_vec(jvm, raw, JMX_MBEAN_ATTRIBUTE_INFO)?;
         let mut attributes = Vec::new();
         for instance in raw {